Revert "utils.bbclass: note for deprecated base_contains"
Recipes from Comcast and the SOC vendors are not generally tested with
OE 2.2 yet and so tend to generate a lot of "base_contains is
deprecated..." warnings. Since these recipes are maintained by others
and using base_contains isn't actually going to cause any real
problems anyway, revert the commit which added the warning.

This reverts commit 1a0e56630c5c27d8899dd0979ae0b86bbe227881.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>