elfutils_0.148: musl build fixes
Note that the patch fixes issues in a number of different elfutils
libraries but is incomplete (there are still build issues in libcpu,
and maybe others). However those issues are hidden since for musl the
recipe now restricts the build to libelf only, ie:

  EXTRA_OEMAKE_libc-musl = "-C libelf"

ie this is a temp solution and should be cleaned up and reviewed
before pushing upstream to meta-gplv2.

For reference:

  https://github.com/GregorR/musl-cross/blob/master/patches/elfutils-0.152-musl.diff
  https://github.com/linuxkit/linuxkit/blob/master/projects/ebpf/ebpf.build/100-musl-compat.patch

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>